+++ src/Mages.Core/Tokens/NumberToken.cs
@@ -33,5 +33,10 @@
         public TextPosition Start => _start;
 
         public TextPosition End => _end;
+
+        public override String ToString()
+        {
+            return $"Number / {_start} -- {_end} / '{_value}'";
+        }
     }
 }