+++ src/Mages.Core/Tokens/StringToken.cs
@@ -30,5 +30,10 @@
         public TextPosition Start => _start;
 
         public TextPosition End => _end;
+
+        public override String ToString()
+        {
+            return $"String / {_start} -- {_end} / '{_content}'";
+        }
     }
 }