+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Added missing helpers and constants (#98)
 - Added complex helpers (#97)
 - Added complex numbers support (#96)
+- Changed the range operator from `:` to `..` (#76)
 - Changed library target to .NET Standard 2 (#68)
 - Changed application to run on .NET Core 3.1 (#68)